Rename HashShuffleReader to ShuffleReader.
